+++ .gitignore
@@ -27,5 +27,6 @@ _site/
 
 # mac
 .DS_Store
+
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
