+++ build.sbt
@@ -1,6 +1,5 @@
 val CatsV = "2.9.0"
 val Fs2V = "3.7.0"
-val MonocleV = "3.1.0"
 val MUnitV = "0.7.29"
 val SourcePosV = "1.1.0"
 
@@ -26,9 +25,9 @@ lazy val commonSettings = Seq(
 
 lazy val dc10 = (project in file("."))
   .settings(name := "dc10")
-  .aggregate(`dc10-core`, `dc10-dsl`, `dc10-io`)
+  .aggregate(`dc10-core`, `dc10-io`, `dc10-scala`)
 
-lazy val `dc10-core` = (project in file("modules/dc10-core"))
+lazy val `dc10-core` = (project in file("modules/core"))
   .settings(
     commonSettings,
     name := "dc10-core",
@@ -39,18 +38,18 @@ lazy val `dc10-core` = (project in file("modules/dc10-core"))
     )
   )
 
-lazy val `dc10-dsl` = (project in file("modules/dc10-dsl"))
-  .dependsOn(`dc10-core`)
-  .settings(
-    commonSettings,
-    name := "dc10-dsl",
-  )
-
-lazy val `dc10-io` = (project in file("modules/dc10-io"))
-  .dependsOn(`dc10-core`)
+lazy val `dc10-io` = (project in file("modules/io"))
+  .dependsOn(`dc10-core`, `dc10-scala`)
   .settings(
     name := "dc10-io",
     libraryDependencies ++= Seq(
       "co.fs2" %% "fs2-io" % Fs2V
     )
+  )
+
+lazy val `dc10-scala` = (project in file("modules/scala"))
+  .dependsOn(`dc10-core`)
+  .settings(
+    commonSettings,
+    name := "dc10-scala",
   )
