+++ modules/core/src/main/scala/dc10/Renderer.scala
@@ -1,0 +1,5 @@
+package dc10
+
+trait Renderer[V, A]:
+  def render(input: List[A]): String
+  def version: V
