+++ modules/core/src/main/scala/dc10/VirtualFile.scala
@@ -1,0 +1,5 @@
+package dc10
+
+import java.nio.file.Path
+
+case class VirtualFile(path: Path, contents: String)
