--- modules/dc10-core/src/main/scala/dc10/compile/Config.scala
@@ -1,12 +1,0 @@
-package dc10.compile
-
-import java.nio.file.Path
-
-trait Config[V]:
-  def target: Path
-
-object Config:
-  
-  given Config["scala-3.3.0"] =
-    new Config["scala-3.3.0"]:
-      def target: Path = Path.of("")
