+++ modules/scala/src/main/scala/dc10/scala/ctx/package.scala
@@ -1,0 +1,13 @@
+package dc10.scala.ctx
+
+import dc10.Compiler
+import dc10.scala.ast.Definition.Statement
+
+type ErrorF[A] = Either[List[Compiler.Error], A]
+
+extension (ctx: List[Statement])
+  def ext(s: Statement): ErrorF[List[Statement]] =
+    namecheck(s).map(ctx :+ _)
+  def namecheck(s: Statement): ErrorF[Statement] =
+    // TODO
+    Right(s)
